% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Run diagnostics}
\usage{
diagnostics(obj)
}
\arguments{
\item{obj}{An object of class \code{vaccine_est} returned by est_ce}
}
\value{
A plot of model diagnostics
}
\description{
TO DO
}
\examples{
print("to do")
}
