% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cox.R
\name{est_cox}
\alias{est_cox}
\title{Estimate CVE/CR using Cox model}
\usage{
est_cox(
  dat,
  t_0,
  cve = T,
  cr = T,
  s_out = seq(from = min(dat$v$s), to = max(dat$v$s), l = 101),
  spline_df = NA,
  edge_ind = F,
  ci_type = "logit",
  placebo_risk_method = "Cox",
  return_extras = F
)
}
\arguments{
\item{dat}{A data object returned by load_data}

\item{t_0}{Time point of interest}

\item{cve}{Boolean. If TRUE, the controlled vaccine efficacy (CVE) curve is
computed.}

\item{cr}{Boolean. If TRUE, the controlled risk (CR) curve is computed.}

\item{s_out}{A numeric vector of s-values (on the biomarker scale) for which
cve(s) and/or cr(s) are computed. Defaults to a grid of 101 points
between the min and max biomarker values.}

\item{spline_df}{TO DO}

\item{edge_ind}{TO DO}

\item{ci_type}{One of c("logit", "truncated", "none").}

\item{placebo_risk_method}{One of c("KM", "Cox"). Method for estimating
overall risk in the placebo group. "KM" computes a Kaplan-Meier estimate
and "Cox" computes an estimate based on a marginalized Cox model survival
curve. Only relevant if cve=TRUE.}

\item{return_extras}{Boolean. If set to TRUE, the following quantities (most
    of which are mainly useful for debugging) are returned: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}}
}
\value{
A list containing the following: \itemize{
    \item{\code{s}: A vector of points, corresponding to s_out}
    \item{\code{est}: asdf}
    \item{\code{se}: asdf}
    \item{\code{ci_lo}: asdf}
    \item{\code{ci_hi}: asdf}
}
}
\description{
Estimate controlled vaccine efficacy (CVE) and/or controlled
    risk (CR) using a marginalized Cox proportional hazards model
}
\examples{
print("to do")
}
\references{
Kenny, A., Gilbert P., and Carone, M. (2023). Inference for
    controlled risk and controlled vaccine efficacy curves using a
    marginalized Cox proportional hazards model

Gilbert P., Fong Y., Kenny A., and Carone, M. (2022). A
    Controlled Effects Approach to Assessing Immune Correlates of Protection.
}
