% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_np.R
\name{est_np}
\alias{est_np}
\title{Estimate CVE/CR nonparametrically}
\usage{
est_np(
  dat,
  cve = T,
  cr = T,
  s_out = seq(from = min(dat$s), to = max(dat$s), l = 101),
  params = list(),
  verbose = F
)
}
\arguments{
\item{dat}{A data object returned by load_data}

\item{cve}{Boolean. If TRUE, the controlled vaccine efficacy (CVE) curve is
computed.}

\item{cr}{Boolean. If TRUE, the controlled risk (CR) curve is computed.}

\item{s_out}{A numeric vector of s-values (on the biomarker scale) for which
cve(s) and/or cr(s) are computed. Defaults to a grid of 101 points
between the min and max biomarker values.}

\item{params}{A list of key value pairs that can be used to select or tune
    various nuisance estimators. See examples. These include: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}}

\item{verbose}{A Boolean. If set to TRUE, intermediate output will be
displayed.}

\item{t_0}{Time point of interest}
}
\value{
A list containing the following: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}
}
\description{
Estimate controlled vaccine efficacy (CVE) and/or controlled
    risk (CR) using the monotone-constrainted nonparametric method of Kenny
    et al. 2023.
}
\examples{
print("to do")
#list(cond_density_type="binning", cond_density_nbins=10)
}
\references{
Kenny, A., Gilbert P., and Carone, M. (2023). Nonparametric
    inference for the controlled risk and controlled vaccine efficacy curves.

Gilbert P., Fong Y., Kenny A., and Carone, M. (2022). A
    Controlled Effects Approach to Assessing Immune Correlates of Protection.
}
