% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sve_utils.R
\name{est_sve}
\alias{est_sve}
\title{Stochastic vaccine efficacy evaluation for immune correlates analyses}
\usage{
est_sve(
  ve_dat,
  Y,
  C = NA_character_,
  S,
  A,
  W,
  B,
  delta_grid = seq(-0.5, 0.5, by = 0.25),
  estimator = c("tmle", "onestep"),
  fluctuation = c("weighted", "standard"),
  weighting = c("identity", "variance"),
  ci_level = 0.95,
  ci_type = c("marginal", "simultaneous"),
  gps_bound = 0.01,
  gps_est = sl3::Lrnr_density_semiparametric$new(mean_learner = sl3::Lrnr_glm_fast$new()),
  or_est = sl3::Lrnr_glm_fast$new(),
  tps_est = sl3::Lrnr_glm_fast$new(),
  eif_est = c("hal", "glm")
)
}
\arguments{
\item{ve_dat}{A rectangular data object (e.g., \code{data.frame}) containing
data on study units from a vaccine efficacy trial. See the arguments below
for details on the variables that ought to be included in \code{ve_dat}.}

\item{Y}{An atomic \code{character} naming a variable in \code{ve_dat} that
corresponds to the outcome of interest (an infection/disease endpoint);
currently, this is limited to binary endpoints. Future extensions include
support for time-to-event endpoints subject to right-censoring.}

\item{C}{An atomic \code{character} naming a variable in \code{ve_dat} that
corresponds to an indicator of whether a study unit was lost to follow-up
during the VE trial. The default of \code{NA_character_} assumes no loss to
follow-up.}

\item{S}{An atomic \code{character} naming a variable in \code{ve_dat} that
corresponds to the measurements of an immunologic marker of interest's
measured activity post-vaccination.}

\item{W}{A vector \code{character} naming those variables in \code{ve_dat}
that correspond to measured baseline covariates that are to comprise the
adjustment set when computing estimates of nuisance parameters.}

\item{B}{A vector \code{character} naming a variable in \code{ve_dat} that
corresponds to an indicator of inclusion in a designed two-phase sample
(e.g., case-control or case-cohort design). This is used to compute inverse
probability of sampling weights to recover population-level inference.}

\item{delta_grid}{A vector \code{numeric} of values by which measurements in
the immune marker specified by \code{S} should be mean-shifted to obtain
counterfactual risk estimates.}

\item{estimator}{The type of estimator to be fit, either \code{"tmle"} for
targeted maximum likelihood estimation or \code{"onestep"} for a one-step
estimator. Both estimators are semiparametric-efficient and compatible with
machine learning-based estimation of nuisance parameters.}

\item{fluctuation}{The method to be used in the submodel fluctuation step
(targeting step) to compute the TML estimator. The choices are "standard"
and "weighted", which correspond to how the auxiliary covariate is included
in the logistic tilting regression model in the TML updating step.}

\item{weighting}{Whether to weight each parameter estimate by the inverse of
its variance (in order to improve stability of the resultant MSM fit) or to
simply weight all parameter estimates equally. The default is the option
\code{"identity"}, weighting all estimates identically.}

\item{ci_level}{A \code{numeric} indicating the desired coverage level of a
Wald-style confidence interval to be computed for the parameter estimate.}

\item{ci_type}{Whether to construct a simultaneous confidence band covering
all parameter estimates at once or marginal confidence intervals covering
each parameter estimate separately. The default is to construct marginal
confidence intervals for each parameter estimate rather than a simultaneous
confidence band.}

\item{gps_bound}{Atomic (\code{numeric}) specifying the lower limit of the
generalized propensity score estimates to be tolerated (with a default of
0.01). Any estimates below this are truncated to max(1/n, this bound).}

\item{gps_est}{...}

\item{or_est}{...}

\item{tps_est}{...}

\item{eif_est}{...}

\item{An}{atomic \code{character} naming a variable in \code{ve_dat} that
corresponds to an indicator of vaccination status or of assignment to the
vaccine versus placebo arm in the VE trial.}
}
\description{
Stochastic vaccine efficacy evaluation for immune correlates analyses
}
