% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall.R
\name{overall}
\alias{overall}
\title{Estimate overall risk and VE using Cox model}
\usage{
overall(dat, t_0, method = "Cox", risk = T, ve = T, ci_type = "logit")
}
\arguments{
\item{dat}{A data object returned by load_data}

\item{t_0}{Time point of interest}

\item{risk}{Boolean. If TRUE, the controlled risk (CR) curve is computed.}

\item{ve}{Boolean. If TRUE, the controlled vaccine efficacy (CVE) curve is
computed.}

\item{ci_type}{TO DO}
}
\value{
A list containing the following: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}
}
\description{
TO DO
}
\examples{
print("to do")
}
