% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_med_cox}
\alias{params_med_cox}
\title{Set parameters controlling Cox model estimation of mediation effects}
\usage{
params_med_cox(edge_ind = F)
}
\arguments{
\item{edge_ind}{Boolean. If TRUE, an indicator variable corresponding to the
lower limit of the marker will be included in the Cox model linear
predictor.}
}
\value{
A list of options.
}
\description{
This should be used in conjunction with \code{\link{est_med}} to
    set parameters controlling Cox model estimation of mediation effects;
    see examples.
}
\examples{
data(hvtn505)
dat <- load_data(time="HIVwk28preunblfu", event="HIVwk28preunbl", vacc="trt",
                 marker="IgG_V2", covariates=c("age","BMI","bhvrisk"),
                 weights="wt", ph2="casecontrol", data=hvtn505)
\donttest{
ests_med <- est_med(
  dat = dat,
  type = "Cox",
  t_0 = 578,
  params_np = params_med_cox(edge_ind=TRUE)
)
}
}
