% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sve_utils.R
\name{plot_sve}
\alias{plot_sve}
\title{Plot of working marginal structural model for stochastic vaccine efficacy}
\usage{
plot_sve(x, ...)
}
\arguments{
\item{x}{Object of class \code{txshift_msm} as produced by a call to
\code{\link{msm_vimshift}}.}

\item{...}{Additional arguments passed to \code{plot} as necessary.}
}
\description{
Plot of working marginal structural model for stochastic vaccine efficacy
}
\details{
Creates a visualization of the intervention-specific counterfactual
 means on the log(relative risk) scale as well as the working marginal
 structural model summarizing the trend across posited values of the
 intervention.
}
