% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Calculate summary statistics}
\usage{
summary_stats(dat, quietly = F)
}
\arguments{
\item{dat}{A data object returned by `load_data`.}

\item{quietly}{Boolean. If true, output will not be printed.}
}
\value{
A list containing the following: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}
}
\description{
TO DO
}
\examples{
print("to do")
}
