% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cox.R
\name{est_cox}
\alias{est_cox}
\title{Estimate CVE/CR using Cox model}
\usage{
est_cox(
  dat,
  t_0,
  cve = T,
  cr = T,
  s_out = seq(from = min(dat$s), to = max(dat$s), l = 101),
  ci_type = "logit",
  grid_size = list(y = 101, s = 101, x = 5),
  return_extras = F,
  verbose = F
)
}
\arguments{
\item{dat}{A data object returned by load_data}

\item{t_0}{Time point of interest}

\item{cve}{Boolean. If TRUE, the controlled vaccine efficacy (CVE) curve is
computed.}

\item{cr}{Boolean. If TRUE, the controlled risk (CR) curve is computed.}

\item{s_out}{A numeric vector of s-values (on the biomarker scale) for which
cve(s) and/or cr(s) are computed. Defaults to a grid of 101 points
between the min and max biomarker values.}

\item{ci_type}{TO DO}

\item{grid_size}{A list containing the following three keys: \itemize{
    \item{\code{y}: grid size for time values}
    \item{\code{s}: grid size for marker values}
    \item{\code{x}: grid size for covariate values}
}
    This controls rounding of data values. Decreasing the grid size values
    results in shorter computation times, and increasing the grid size values
    results in more precise estimates. If grid_size$s=101, this means that a
    grid of 101 equally-spaced points (defining 100 intervals) will be
    created from min(S) to max(S), and each S value will be rounded to the
    nearest grid point. For grid_size$y, a grid will be created from 0 to
    t_0, and then extended to max(Y). For grid_size$x, a separate grid is
    created for each covariate column (binary/categorical covariates are
    ignored).}

\item{return_extras}{Boolean. If set to TRUE, the following quantities (most
    of which are mainly useful for debugging) are returned: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}}

\item{verbose}{A Boolean. If set to TRUE, intermediate output will be
displayed.}
}
\value{
A list containing the following: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}
}
\description{
Estimate controlled vaccine efficacy (CVE) and/or controlled
    risk (CR) using a marginalized Cox proportional hazards model
}
\examples{
print("to do")
}
\references{
Kenny, A., Gilbert P., and Carone, M. (2023). Inference for
    controlled risk and controlled vaccine efficacy curves using a
    marginalized Cox proportional hazards model

Gilbert P., Fong Y., Kenny A., and Carone, M. (2022). A
    Controlled Effects Approach to Assessing Immune Correlates of Protection.
}
