% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_np.R
\name{est_np}
\alias{est_np}
\title{Estimate CVE/CR nonparametrically}
\usage{
est_np(
  dat,
  t_0,
  cve = T,
  cr = T,
  s_out = seq(from = min(dat$s), to = max(dat$s), l = 101),
  ci_type = "logit",
  cf_folds = 1,
  edge_corr = "none",
  params = list(),
  grid_size = list(y = 100, s = 100, x = 10),
  verbose = F
)
}
\arguments{
\item{dat}{A data object returned by load_data}

\item{t_0}{Time point of interest}

\item{cve}{Boolean. If TRUE, the controlled vaccine efficacy (CVE) curve is
computed.}

\item{cr}{Boolean. If TRUE, the controlled risk (CR) curve is computed.}

\item{s_out}{A numeric vector of s-values (on the biomarker scale) for which
cve(s) and/or cr(s) are computed. Defaults to a grid of 101 points
between the min and max biomarker values.}

\item{ci_type}{TO DO}

\item{cf_folds}{TO DO}

\item{edge_corr}{TO DO}

\item{params}{A list of key value pairs that can be used to select or tune
    various nuisance estimators. See examples. These include: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}}

\item{grid_size}{A list containing the following three keys:
    \itemize{
    \item{\code{y}: grid size for time values}
    \item{\code{s}: grid size for marker values}
    \item{\code{x}: grid size for covariate values}
}
    This controls rounding of covariates, which results in substantially
    shorter computation times. If grid_size$y=100, this means that a grid of
    equally-spaced points will be created from min(dat$y) to max(dat$y), and
    each dat$y value will be rounded to the nearest grid point. The same
    holds for grid_size$s. For grid_size$x, a separate grid is created for
    each covariate column (binary and categorical covariates are ignored).}

\item{verbose}{A Boolean. If set to TRUE, intermediate output will be
displayed.}
}
\value{
A list containing the following: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}
}
\description{
Estimate controlled vaccine efficacy (CVE) and/or controlled
    risk (CR) using the monotone-constrainted nonparametric method of Kenny
    et al. 2023.
}
\examples{
print("to do")
#list(cond_density_type="binning", cond_density_nbins=10)
}
\references{
Kenny, A., Gilbert P., and Carone, M. (2023). Nonparametric
    inference for the controlled risk and controlled vaccine efficacy curves.

Gilbert P., Fong Y., Kenny A., and Carone, M. (2022). A
    Controlled Effects Approach to Assessing Immune Correlates of Protection.
}
