% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_np.R
\name{est_np}
\alias{est_np}
\title{Estimate CVE/CR nonparametrically}
\usage{
est_np(
  dat,
  t_0,
  cr = T,
  cve = T,
  s_out = seq(from = min(dat$v$s, na.rm = T), to = max(dat$v$s, na.rm = T), l = 101),
  ci_type = "logit",
  placebo_risk_method = "KM",
  cf_folds = 1,
  edge_corr = F,
  params = list(),
  grid_size = list(y = 101, s = 101, x = 5),
  return_extras = F
)
}
\arguments{
\item{dat}{A data object returned by load_data}

\item{t_0}{Time point of interest}

\item{cr}{Boolean. If TRUE, the controlled risk (CR) curve is computed.}

\item{cve}{Boolean. If TRUE, the controlled vaccine efficacy (CVE) curve is
computed.}

\item{s_out}{A numeric vector of s-values (on the biomarker scale) for which
cve(s) and/or cr(s) are computed. Defaults to a grid of 101 points
between the min and max biomarker values.}

\item{ci_type}{TO DO}

\item{placebo_risk_method}{One of c("KM", "Cox"). Method for estimating
overall risk in the placebo group. "KM" computes a Kaplan-Meier estimate
and "Cox" computes an estimate based on a marginalized Cox model survival
curve. Only relevant if cve=TRUE.}

\item{cf_folds}{An integer representing the number of cross-fitting folds to
use. If cf_folds=1, cross-fitting will not be done.}

\item{edge_corr}{Boolean. If TRUE, the edge correction is performed. It is
not recommended that the edge correction is performed unless there are
at least 10 events corresponding to the marker lower limit}

\item{params}{A list of key value pairs that can be used to select or tune
    various nuisance estimators. See examples. These include: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}}

\item{grid_size}{A list containing the following three keys: \itemize{
    \item{\code{y}: grid size for time values}
    \item{\code{s}: grid size for marker values}
    \item{\code{x}: grid size for covariate values}
}
    This controls rounding of data values. Decreasing the grid size values
    results in shorter computation times, and increasing the grid size values
    results in more precise estimates. If grid_size$s=101, this means that a
    grid of 101 equally-spaced points (defining 100 intervals) will be
    created from min(S) to max(S), and each S value will be rounded to the
    nearest grid point. For grid_size$y, a grid will be created from 0 to
    t_0, and then extended to max(Y). For grid_size$x, a separate grid is
    created for each covariate column (binary/categorical covariates are
    ignored).}

\item{return_extras}{Boolean. If set to TRUE, the following quantities (most
    of which are mainly useful for debugging) are returned: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}}
}
\value{
A list containing the following: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}
}
\description{
Estimate controlled vaccine efficacy (CVE) and/or controlled
    risk (CR) using the monotone-constrainted nonparametric method of Kenny
    et al. 2023.
}
\note{
- This method assumes that risk decreases as the biomarker increases. If it
    is assumed that risk increases as the biomarker increases, reverse the
    scale of the biomarker.
}
\examples{
print("to do")
}
\references{
Kenny, A., Gilbert P., and Carone, M. (2023). Nonparametric
    inference for the controlled risk and controlled vaccine efficacy curves.

Gilbert P., Fong Y., Kenny A., and Carone, M. (2022). A
    Controlled Effects Approach to Assessing Immune Correlates of Protection.
}
