% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load and format data object}
\usage{
load_data(time, event, marker, covariates, weights, ph2)
}
\arguments{
\item{time}{A numeric vector of observed event or censoring times.}

\item{event}{A vector of binary values corresponding to whether the observed
time represents an event time (1) or a censoring time (0). Accepts either
integer (0/1) or Boolean (T/F) values.}

\item{marker}{A numeric vector of biomarker values.}

\item{covariates}{A dataframe. Columns values should be either numeric,
binary, or factors. Character columns will be converted into factors.}

\item{weights}{A numeric vector of inverse-probability-of-sampling (IPS)
weights.}

\item{ph2}{A vector of binary values corresponding to whether the individual
is in the phase-two cohort (1) or not (0). Accepts either integer (0/1)
or Boolean (T/F) values.}

\item{vacc}{A vector of binary values corresponding to whether the individual
is in the vaccine group (1) or the placebo group (0). Accepts either
integer (0/1) or Boolean (T/F) values.}
}
\value{
A list containing the following: \itemize{
    \item{\code{one}: asdf}
    \item{\code{two}: asdf}
    \item{\code{three}: asdf}
}
}
\description{
This function takes in user-supplied data and returns a data
    object that can be read in by \code{est_np()}, \code{est_cox()},
    and \code{hyptest_np()}. Data is expected to come from a case-cohort
    sampling design and be filtered to include all phase-one individuals.
}
\examples{
print("to do")
}
