% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load and format data object}
\usage{
load_data(
  time,
  event,
  vacc,
  marker,
  covariates,
  weights,
  ph2,
  strata = NA,
  data
)
}
\arguments{
\item{time}{The name of the numeric variable representing observed event or
censoring times.}

\item{event}{The name of the binary variable corresponding to whether the
observed time represents an event time (1) or a censoring time (0).
Either integer (0/1) or Boolean (T/F) values are allowed.}

\item{vacc}{A vector of binary values corresponding to whether the individual
is in the vaccine group (1) or the placebo group (0). Accepts either
integer (0/1) or Boolean (T/F) values.}

\item{marker}{A numeric vector of biomarker values.}

\item{covariates}{A dataframe. Columns values should be either numeric,
binary, or factors. Character columns will be converted into factors.}

\item{weights}{A numeric vector of inverse-probability-of-sampling (IPS)
weights.}

\item{ph2}{A vector of binary values corresponding to whether the individual
is in the phase-two cohort (1) or not (0). Accepts either integer (0/1)
or Boolean (T/F) values.}

\item{strata}{A vector of strata identifiers (for two-phase sampling strata).
weights.}
}
\value{
An object of class \code{vaccine_dat}.
}
\description{
This function takes in user-supplied data and returns a data
    object that can be read in by \code{est_np()}, \code{est_cox()}, and
    other estimation functions. Data is expected to come from a two-phase
    sampling design and should be filtered to include all phase-one
    individuals.
}
\examples{
data(hvtn505)
dat <- load_data(time="HIVwk28preunblfu", event="HIVwk28preunbl", vacc="trt",
                 marker="logpctpos_scaled", covariates=c("age","BMI","bhvrisk"),
                 weights="wt", ph2="casecontrol", data=hvtn505)
}
