% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Calculate summary statistics}
\usage{
summary_stats(dat, quietly = F)
}
\arguments{
\item{dat}{A data object returned by `load_data`.}

\item{quietly}{Boolean. If true, output will not be printed.}
}
\value{
A list containing values of various summary statistics.
}
\description{
TO DO
}
\examples{
data(hvtn505)
dat <- load_data(time="HIVwk28preunblfu", event="HIVwk28preunbl", vacc="trt",
                 marker="IgG_V2", covariates=c("age","BMI","bhvrisk"),
                 weights="wt", ph2="casecontrol", data=hvtn505)
summary_stats(dat=dat)
}
