% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine-package.R
\docType{package}
\name{vaccine-package}
\alias{vaccine}
\alias{vaccine-package}
\title{vaccine: Statistical tools for immune correlates analysis of vaccine clinical trial data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains various semiparametric and nonparametric statistical tools for immune correlates analysis of vaccine clinical trial data. This includes calculation of summary statistics and estimation of risk, vaccine efficacy, controlled risk, and controlled vaccine efficacy.
}
\author{
\strong{Maintainer}: Avi Kenny \email{avi.kenny@gmail.com}

}
\keyword{internal}
